+++ ocbpy/boundaries/__init__.py
@@ -10,12 +10,10 @@ files          Boundary file utilities
 dmsp_ssj_files DMSP SSJ boundary file utilities
 
 """
-from __future__ import absolute_import
-import logging
-
+from ocbpy import logger
 from ocbpy.boundaries import files
 
 try:
     from ocbpy.boundaries import dmsp_ssj_files
 except ImportError as ierr:
-    logging.warning(ierr)
+    logger.warning(ierr)