+++ ocbpy/instruments/__init__.py
@@ -12,9 +12,7 @@ general     General file loading and testing routines
 pysat       General pysat Instrument loading routines: https://github.com/pysat
 
 """
-from __future__ import (absolute_import)
-import logging
-
+from ocbpy import logger
 from ocbpy.instruments import general
 from ocbpy.instruments.general import test_file
 from ocbpy.instruments import supermag
@@ -23,4 +21,4 @@ from ocbpy.instruments import vort
 try:
     from ocbpy.instruments import pysat_instruments
 except ImportError as ierr:
-    logging.warning(ierr)
+    logger.warning(ierr)