+++ Changelog.rst
@@ -3,6 +3,13 @@ Changelog
 
 Summary of all changes made since the first stable release
 
+0.2.X (XX-XX-2021)
+------------------
+* DOC: Improved the PEP8 compliance in the documentation examples
+* BUG: Fixed header initialization error general instrument loading routine
+* MAINT: Removed support for Python 2.7 and 3.5
+
+
 0.2.1 (11-24-2020)
 ------------------
 * DOC: Updated examples in README